mgr/cephadm: add 'haproxy' service type
This is a slightly more opinionated version of the haproxy half of the
HA_RGW service.

- specify which backend service we will proxy (there may be multiple RGW
  services/zones in the cluster)
- simplify option names (remove ha_proxy_ prefix, remove unneeded options,
  require only a handful and use defaults for the others)
- autogenerate a password if one is not provided (it can be fetched from
  config-key)

Note that this will subtlely break the ha-rgw service because haproxy
daemons are now assumed to belong to an haproxy service, and we don't
bother to sort things out since ha-rgw will be removed soon.

Signed-off-by: Sage Weil <sage@newdream.net>