librbd: fix qos m_io_throttled assertion failure
I found that the operation of ++m_io_throttled and --m_io_throttled is
done in the main thread and the timer thread respectively. Although
m_io_throttled is an atomic variable, ++ The operation of
m_io_throttled in the main thread may be slower than --m_io_throttled
in the timer thread, so a coredump will appear when
ceph_assert(m_io_throttled.load()> 0) is performed The code is as
follows:

if (needs_throttle(peek_item)) {
 ldout(cct, 15) << "throttling IO " << peek_item << dendl;

 ++m_io_throttled;
 // dequeue the throttled item
 ThreadPool::PointerWQ<ImageDispatchSpec<I> >::_void_dequeue();
 return nullptr;
}

void ImageRequestWQ::handle_throttle_ready(int r, ImageDispatchSpec
 *item, uint64_t flag) {
 CephContext *cct = m_image_ctx.cct;
 ldout(cct, 15) << "r=" << r << ", " << "req=" << item << dendl;
 ceph_assert(m_io_throttled.load() > 0);
 item->set_throttled(flag);
 if (item->were_all_throttled()) {
  this->requeue_back(item);
  --m_io_throttled;
  this->signal();
 }
}

Putting ceph_assert behind requeue_back can guarantee ++m_io_throttled
occuring first ，This is because _void_dequeue and requeue_back need to
use the same lock and _void_dequeue hold the lock.

this issue only exists in octopus and nautilus, so it's not
cherry-picked from master.

Fixes :https://tracker.ceph.com/issues/49949
Signed-off-by: pkulijiawei <lijiawei1@chinatelecom.cn>
Signed-off-by: wuxuehan <wuxuehan@chinatelecom.cn>