cephadm: only bootstrap using image that matches cephadm version
Only allow bootstrap to deploy if the cephadm version matches the
ceph version in the container.  Allow the master branch version of cephadm
to deploy the latest stable version as well (at least for now).

Provide a flag to force bootstrap to continue despite the check.

Move the _pull_image call up into bootstrap so that it is easier to see
when it happens.

Fixes: https://tracker.ceph.com/issues/49884
Signed-off-by: Sage Weil <sage@newdream.net>
(cherry picked from commit 1b5a068ddfc84002cefb384ea01e97b5300408a1)

# Conflicts:
#	src/cephadm/cephadm
 - default images are diff for pacific branch