librbd/cache/pwl: Discard write back cache in removing image request
In the remove operation, discard the write back cache and skip the PWL
initialization. so that the remove command can succeed and clean up the
cache no matter what the cache state is in.
Scenario: if the PWL directory is deleted before the remove operation,
ensure that the remove operation will not fail.

Signed-off-by: Yin Congmin <congmin.yin@intel.com>