common/pick_address: Allow binding on loopback iface
in bfff08957190c8140c0f444a26e40eeef3c2a57f, "lo" is also skipped when
daemon is trying to find an address to bind. but that change reverts the
fix of 176c8ab57889c97a9355564239b1c5f6fd440772, to address the problem.

an option named "ms_bind_exclude_lo_iface" is added, it defaults to
"true". but it can be changed to false to allow daemon to bind on "lo".

Fixes: https://tracker.ceph.com/issues/50012
Signed-off-by: Kefu Chai <kchai@redhat.com>