client: only check pool permissions for regular files
There is no need to do a check_pool_perm() on anything that isn't
a regular file, as the MDS is what handles talking to the OSD in
those cases. Just return 0 if it's not a regular file.

Signed-off-by: Xiubo Li <xiubli@redhat.com>
(cherry picked from commit 59a3006b88f479cb5333e16fe30201ea14ab1717)