test: disable mgr/mirroring for `test_mirroring_init_failure_with_recovery` test
The test simulates failures by enabling mirroring for file system
via mon interface. Doing this does not create `cephfs-mirror` index
obejct that trips the mirror daemon to report failed status via
asok interface -- so that the test can verify this.

However, mgr/mirroring module gets to know that mirroring is
enabled for this file system (via FSMap) and tried to load the
directory map, but fails to do so and throws a backtrace in the
logs. Fix this by disabling mgr/mirroring before mirroring is
enabled via mon command.

Signed-off-by: Venky Shankar <vshankar@redhat.com>