mgr/dashboard: do not import tools in access_control
this addresses a regression introduced by
2cd94293268116838c3ddcebdedde4fbd9cb93aa. which

from ..tools import ensure_str

and it causes recursive import.

so, in this change, an copy of ensure_str() is added
to access_control.py

this change is not cherry-picked from master, as the offending commit
which is fixed by this change is not included in master.

Signed-off-by: Kefu Chai <kchai@redhat.com>