rgwlc: remove explicit lc shard resets at start-of-run
This is an alternative solution to the (newly exposed) lifecycle
shard starvation problem reported in b y Jeegen Chen.

There was always an implied starvation condition implied by the
reset of lc shard head at the start of processing.  The introduction
of "stale sessions" in parallel lifecycle changes made it more
visible, in particular when rgw_lc_debug_interval was set to a small
value and many buckets had lifecycle policy.

My hypothesis in this change is that lifecycle processing for each
lc shard should /always/ continue through the full set of eligible
buckets for the shard, regardless of how many processing cycles might
be required to do so.  There is no semantic basis for restarting at the
first eligible bucket in any circumstance, and doing so was itself
a latent bug.

Fixes: https://tracker.ceph.com/issues/49446

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>