osd: drop the legacy, pingless Watch support.
`CEPH_OSD_WATCH_OP_LEGACY_WATCH` is long dead but it still
makes the detection of `ms_handle_reset()` crucial for `Watch`.
It's because the reset event is used to activate the watch's
timeout timer -- see the `Watch::disconnect()` method.

Thus, killing the legacy op has a nice consequence: it allows
for eradication of the `WatchConState`; the reset detection
isn't important for modern variant of the op as it relies on
our own, ping-based mechanism.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>