rbd-nbd: add ability to restore and attach the nbd-devices
Problem:
-------
Currently, with the rbd-nbd tool, we can map the nbd devices and scale
them as needed. But in case if the node is rebooted, there is no way to restore
the nbd device mappings and one has to do remap of the big list of devices
manually. This is a serious problem in container environments, if the container
hosting the rbd-nbd processes is restarted, all the mapping is just gone and
The application will start seeing the IO errors.

Solution:
--------
With these changes, on every successful Map operation, we save the
details of the nbd device mappings to a persistent file (here on called as
saveconfig file) in the JSON format which will be used for restore. Similarly,
on Unmap we clean the JSON object related to the device.

Thanks to the attach command, the restore command will leverage it and
brings the rbd-nbd processes to life.

The default path of the JSON file is /etc/ceph/saveconfig.json

Example usage:
-------------
$ rbd-nbd list-mapped
id      pool           namespace  image    snap  device
617318  rbd-pool                  image1   -     /dev/nbd0
617356  rbd-pool                  image2   -     /dev/nbd1
617394  rbd-pool                  image3   -     /dev/nbd2
617433  rbd-pool                  image4   -     /dev/nbd3
617472  rbd-pool                  image5   -     /dev/nbd4

$ pkill -9 rbd-nbd

$ rbd-nbd restore
=> Attempting to attach: /dev/nbd0 -> rbd-pool/image1
SUCCESS
=> Attempting to attach: /dev/nbd1 -> rbd-pool/image2
SUCCESS
=> Attempting to attach: /dev/nbd2 -> rbd-pool/image3
SUCCESS
=> Attempting to attach: /dev/nbd3 -> rbd-pool/image4
SUCCESS
=> Attempting to attach: /dev/nbd4 -> rbd-pool/image5
SUCCESS

Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>