rbd-nbd: option to provide configurable saveconfig file
Environments like containers running rbd-nbd in the pods, need to tune
this saveconfig path to the persistent host-path (bind mounted) locations.
Not just that, for example, CSI drivers need a saveconfig file per rbd-nbd
device.

To satisfy the above requirements, changes here introduce --saveconfig-file
option with the map/restore/unmap commands, which allows tuning the default
location of the saveconfig file.

Please note its the duty of the user to provide the same path of the
saveconfig file at the time of unmap if the rbd-nbd process is attached
manually instead of invoking it via the restore command, especially if
the --saveconfig-file option is not used with attach command.

Example usage:
-------------
$ ./bin/rbd-nbd map --saveconfig-file /tmp/rbd-nbd-cfg.json
$ ./bin/rbd-nbd restore --saveconfig-file /tmp/rbd-nbd-cfg.json
$ ./bin/rbd-nbd unmap

Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>