doc/_themes: remove spacing after `ul li p`
in the latest document generated from RtD, the spacing after `ul li p`
elements is set to 24px as the plain `p` elements. but this the lists
more sparse and difficult to read.

in this change, the spacing is restored to 0 as it was in old theme.css
in sphinx_rtd_theme.

Signed-off-by: Kefu Chai <kchai@redhat.com>