mgr/cephadm: ingress: add optional virtual_interface_networks
It may be that the virtual IP we want to use is not in the same network
as any existing IPs on the host.  In that case, allow the spec to specify
a list of networks to match against existing IPs so that a match will
identify an ethernet interface to use.

Signed-off-by: Sage Weil <sage@newdream.net>
(cherry picked from commit bbf6a12752092b406abbec1e600533366ac59548)