cmake: update civetweb.h on demand
instead of using a dedicated target for updating civetweb.h, use
add_custom_command() with OUTPUT option.

so the generated rule does not rerun the command specified by the
civetweb_h target every time we build the targets depending on it.

use the header file as the dependency helps to improve the readability
as there is one less link in the dependency chain, and by specifying the
OUTPUT, cmake is able to figure out the dependency on the header file so
it does not try to regenerate it every time.

Signed-off-by: Kefu Chai <kchai@redhat.com>