common: make y2c.py choke on duplicate keys
Commit c3bbb05deb3b ("common: generate legacy_config_opts.h from
.yaml.in files") inadvertently reverted a change of a default value by
adding a second "default" key with the old value.  This was corrected
in commit 041d749dba16 ("common/options/global: correct default of
auth_mon_ticket_ttl"), but highlights that mis-merging a yaml file
is rather easy.

To prevent this happening again, fail the build if duplicate keys
exist in any of src/common/options/*.yaml.in files.

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>