cephadm: distribute client.admin keyring+conf to label:_admin on bootstrap
If we are placing ceph.conf in /etc/ceph (the default), tell the cluster
to continue doing this going forward to hosts with the '_admin' label.

This doesn't induce the user to add the admin label to other hosts too,
unfortunately--e probably want them to add the admin label to other mons,
for instance--but it is a start.

Signed-off-by: Sage Weil <sage@newdream.net>
(cherry picked from commit 08d21ccc949aa19c08c7e806a5b58c0180f5cbe2)