mgr/cephadm/schedule: assign/map ranks
If we are passed a rank_map, use it maintain one daemon per rank, where
the ranks are consecutive non-negative integers starting from 0.

A bit of refactoring in place() so that we only do the rank allocations
on slots we are going to use (no more than count).

Signed-off-by: Sage Weil <sage@newdream.net>
(cherry picked from commit de03942f806cb16e00259f2397446afb9322c3f1)