doc/_ext/ceph_confval: print "false" default values
before this change, we check the exisitence of default value of an
option using "if default", but this is wrong. what if "default" is an
empty string or a numeric value of 0, or just False?

after this change, its exisitence is checked using "if default is
defined", so all "false" default values are rendered as well.

Signed-off-by: Kefu Chai <kchai@redhat.com>