pybind/mirroring: start from correct state during policy init
So, there were couple of issues here:

First, Policy::init() checks if the directory path has a purge
operation (remove) requested. This check is incorrect since the
`purging` flag is fetched from the directory mapping rather than
from the directory entry in the mapping itself.

Second, the starting state for a directory path was being set
again. Though this was not incorrect, it is redundant and can
be removed.

Signed-off-by: Venky Shankar <vshankar@redhat.com>
(cherry picked from commit 8886eb8389f3c4e52760930f545383fe28f81859)