doc/cephadm: rewrite "deploy. mon. w/cephadm"
This PR corrects various minor usage errors in
the subsection "Deploying Monitoring with cephadm"
of the "monitoring" section of the cephadm guide.

It also adds prompts, and cleans up the list rst a
bit.

This is the first of two PRs planned for this section.
The second PR will create a sub-subsection that explains
an additional installation step that is necessary when
a Ceph cluster and a user accessing that cluster are in
different DNS zones. But that's another story, for another
PR.

Signed-off-by: Zac Dover <zac.dover@gmail.com>