[mgr/dashboard] build fails if python3-virtualenv-xx is not installed
In particular it fails late, for reasons that are hard to diagnose.

It would probably fail earlier, and be easier to diagnose,
if .../src/tools/setup-virtualenv.sh used `set -e`

Found while trying to build on rhel9-beta which currently does not
have python3-virtualenv-xx available, e.g. in its Code Ready Builder
repo or the BaseOS repo.

(N.B. python3-virtualenv-xx and its dependencies are built in brew
for rhel9, but at the time of this writing aren't in any package repos
for rhel9)

Fixes: https://tracker.ceph.com/issues/50760

Signed-off-by: Kaleb S KEITHLEY <kkeithle@redhat.com>