osd/ECBackend: don't remove in-flight backfill when missing is not primary
This is a direct commit to nautilus branch to recreate the
behaviour we have post-nautilus, after refactoring made in
8a8947d2a32d6390cb17099398e7f2212660c9a1, which seems to be
correct.

The in-flight backfill prevents updating of "backfill complete"
position, which remains on the object before the missing oid. So
when pg is re-peered it retries backfill from this position
instead of entering the clean state.

Fixes: https://tracker.ceph.com/issues/50747
Signed-off-by: Mykola Golub <mgolub@suse.com>