common/options: extract mgr and mon options out
the purposes of this change are:

* faster compilation, so the cmake generator can process .yaml.in
  files in parallel.
* allow daemons to include a subset of options which it is
  interested in.
* better maintainability. by grouping options in different .yaml.in
  files, developers understand who are the consumers of an option.

in this change, options only read by mgr are extracted into mgr.yaml.in,
and options only read by mon are extracted into mon.yaml.in.

so all options in mgr.yaml.in should have "services: mgr" in their
definition by default. the ones in mon.yaml.in have "services: mon".

in the case where options are consumed by multiple services or tools,
the option should add "common" to its "services" if it is supposed to be
consumed by a tool, or "mon" if it is read by monitor as well. in
future, when preprocessing the .yaml.in, we have to look up in other
.yaml files for options whose services includes "mon", when compiling
ceph-mon executable in addition to collecting the options in
mon.yaml.in.

but it takes time to audit all the options, so only part of them are
processed.

Signed-off-by: Kefu Chai <kchai@redhat.com>