mds: remove 'fs_name' from MDSRank
There isn't a need to store a file system's name in a MDSRank
object. The MDSRank has a pointer to a MDSMap object, which already
stores the name.

Also, there isn't a need to pass the file system name in the MMDSMap
message. It should be sufficient to pass the MDSMap in the MMDSMap
message as the the file system name is stored in the MDSMap. Pass a
empty string as map_fs_name in the MMDSMap message. It is  simpler than
removing map_fs_name from the message payload altogether.

Fixes: https://tracker.ceph.com/issues/50852
Signed-off-by: Ramana Raja <rraja@redhat.com>