mon/MonClient: send commands to mon after connected
MonClient sends the pending commands to monitor once it completes the
authentication with the monitor, but the connection is not fully
established by then. it's still in its AUTH_CONNECTING_SIGN phase, and
the server has not sent its server identity to MonClient yet. this does
not happen before the connection enters READY phase.

so if there is any message sent via MonClient before the connection is
"READY", it is encoded with feature of "0", even if connection between
the connected monitor and MonClient has non-zero feature bits. and
that's the case in general.

"fortunately", we are using messenger->ms_can_fast_dispatch(m) to tell
if we can "fast_prepare" a message. and the only "can_fast_dispatch"
message sent to monitor is MMonCommand, whose encoder does not use
features at all. that's why the existing implementation works. but
this is not quite correct though, as the criteria here should be:

- if we can encode the message in the critical path. as send_message()
  is supposed to return immediately. if encoding a message takes lots
  of CPU cycles we should defer this to the messager's worker thread.
- if we can re-encode the message. for instance, if a message encodes
  its data in a destructive way, we are not allowed to re-encode it.

instead of if the message can be processed in the fast dispatch path
on the receiver side. this is addressed in another change. see
https://github.com/ceph/ceph/pull/34659

Signed-off-by: Kefu Chai <kchai@redhat.com>